+++ src/main/java/mediathek/filmlisten/reader/FilmListReader.java
@@ -398,7 +398,13 @@ public class FilmListReader implements AutoCloseable {
             } else
                 processFromFile(source, listeFilme);
 
-            convertTagesschau24(listeFilme);
+            if (SenderFilmlistLoadApprover.isApproved("tagesschau24")) {
+                logger.trace("converting old tagesschau24 entries to sender tagesschau24");
+                convertTagesschau24(listeFilme);
+            }
+            else {
+                logger.trace("conversion of tagesschau24 stopped as it is not approved");
+            }
         } catch (MalformedURLException | URISyntaxException ex) {
             logger.warn(ex);
         }